Remove needless exposing Proof and ProofMessage in shared
